refactor(KlaviyoSwift): relocate event publish off KlaviyoInternal (MAGE-832)
